Corrigindo conflitos
